Move hooks into /src/hooks, clean up useFocusWithin
