+++ apps/web/src/components/Basenames/RegistrationForm.tsx
@@ -2,11 +2,8 @@ import { MinusIcon, PlusIcon } from '@heroicons/react/16/solid';
 import { ConnectButton, useConnectModal } from '@rainbow-me/rainbowkit';
 import { Button, ButtonSizes, ButtonVariants } from 'apps/web/src/components/Button/Button';
 import { Icon } from 'apps/web/src/components/Icon/Icon';
-import {
-  useDiscountedNameRegistrationPrice,
-  useNameRegistrationPrice,
-} from 'apps/web/src/utils/hooks/useNameRegistrationPrice';
-import { useRegisterNameCallback } from 'apps/web/src/utils/hooks/useRegisterNameCallback';
+import { useNameRegistrationPrice } from 'apps/web/src/hooks/useNameRegistrationPrice';
+import { useRegisterNameCallback } from 'apps/web/src/hooks/useRegisterNameCallback';
 import { useCallback, useState } from 'react';
 import { formatEther } from 'viem';
 import { base, baseSepolia } from 'viem/chains';
