+++ apps/web/src/hooks/useIsNameAvailable.ts
