+++ apps/web/src/hooks/useNameRegistrationPrice.ts
